src/cloud_functions/requirements.txt
google-cloud-bigquery>=3.3.5
functions-framework>=3.0.0